refactor: Extract site details dialog and fix flutter controller path
- Extract SiteDetailsDialog from NearbysSitesScreen to separate widget file
- Clean up imports and remove unused code
- Fix flutter_controller_enhanced to use relative path instead of hardcoded absolute path
- Improve code organization and maintainability
